requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.6
PyPDF2==3.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
graphviz==0.20.1
pdfplumber==0.10.3
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
mangum==0.17.0 